Fix category bug with uuid mapping.
Had a bug which wasn't caught on my local unit tests but was on the
production ones (wierd) that the uuid mapping category was wrong.  It
was actually a pretty big flaw which is why I'm glad the unit test
caught it.  However the tests were in the wrong place since I moved
around where the mapping is done.  I created a new test for the events
subscriber, fixed the category problem for the mapping, and added an
additional test for the social history mapping which we didn't have.