Make service classes easier to use
Made it so the fhir search classes will convert strings and numbers into
an exact string match.  This makes it so that search methods can pass in
simple array values without having to create objects each time.  This
was a request from @sjpadgett to make the classes a little easier to
use.