Encounter Appointment enhancements
Added the user and group data elements to the EncounterService so we can
specify what user and group is creating the encounter form.

Changed the rest API to use the user and group from the session for the encounter.

Enhanced the appointment service to return an address as a string.  Useful when displaying addresses in a <pre></pre> tag or w/ the <address></address> tag.

Added a number of helper methods to AppointmentService
getCalendarCategories()
isCheckInStatus()
isCheckOutStatus()
updateAppointmentStatus()
getEncounterForAppointment()
createEncounterForAppointment()

The last two methods will go off the current date of Y-m-d which is not super precise but we have no other linking mechanism right now and we currently do that in the auto-create appointment piece in the appointments service.

Added to the EncountersService
getEncountersForPatientByPid()
getPatientEncounterListWithCategories()

The getPatientEncounterListWithCategories method returns the needed data format that is used by the frontend knockout.js to populate the encounter list dropdown.  This is useful in any kind of frontend rendering / AJAX application for updating that list (using left_nav.setPatientEncounter)

Fixed the getUser method in UserService to return the stringified uuid.