Appointment status tracker, FlowBoard module
Extracted majority of patient tracker methods into PatientTrackerService
so I could call the patient tracker.  Almost went the whole way of
extracting patient tracker but it ended up being a much bigger change
and I had to back out and leave the fetch appointments method alone for
now.

Created the initial underpinnings for extracting the FlowBoard into its
own module and put the drug testing piece into the module as an event
subscriber.  We had a cyclic dependency going o between the tracker and
appointments and extracting it into an event in the subscribe resolved
the dependency.