Encounter Appointment enhancements (#4719)
* Encounter Appointment enhancements

Added the user and group data elements to the EncounterService so we can
specify what user and group is creating the encounter form.

Changed the rest API to use the user and group from the session for the encounter.

Enhanced the appointment service to return an address as a string.  Useful when displaying addresses in a <pre></pre> tag or w/ the <address></address> tag.

Added a number of helper methods to AppointmentService
getCalendarCategories()
isCheckInStatus()
isCheckOutStatus()
updateAppointmentStatus()
getEncounterForAppointment()
createEncounterForAppointment()

The last two methods will go off the current date of Y-m-d which is not super precise but we have no other linking mechanism right now and we currently do that in the auto-create appointment piece in the appointments service.

Added to the EncountersService
getEncountersForPatientByPid()
getPatientEncounterListWithCategories()

The getPatientEncounterListWithCategories method returns the needed data format that is used by the frontend knockout.js to populate the encounter list dropdown.  This is useful in any kind of frontend rendering / AJAX application for updating that list (using left_nav.setPatientEncounter)

Fixed the getUser method in UserService to return the stringified uuid.

* Appointment status tracker, FlowBoard module

Extracted majority of patient tracker methods into PatientTrackerService
so I could call the patient tracker.  Almost went the whole way of
extracting patient tracker but it ended up being a much bigger change
and I had to back out and leave the fetch appointments method alone for
now.

Created the initial underpinnings for extracting the FlowBoard into its
own module and put the drug testing piece into the module as an event
subscriber.  We had a cyclic dependency going o between the tracker and
appointments and extracting it into an event in the subscribe resolved
the dependency.

* Put back missing method.

Was missing a method for the flowboard refactor.  So strange it didn't
throw a php error and just gave a 500 response.

* Filter out core modules from the system.