Fix #4725 user events, uuid generation
Fixes #4725
Made it so the create user event only fires if the user successfully
saves.  Also made it so creating a user generates a uuid for it so we
have it at the time of the save user event.  Added some helper methods
to the user created and user updated event to make them more useful.