Openemr fix 4723 module install script uses OpenEMR sql upgrade service (#4724)
* Fix module semicolon install bug for custom module

Closes #4723. I migrate the logic in the sql_upgrade_fx file into a
SQLUpgradeService object that can be reused in other places.  I change
up the InstallerController for both the upgrade sql and the install sql
process to use the SQLUpgradeServiceObject.

I also fix a bug in the uninstall custom module that was throwing an
error due to a recent laminas pdo change with the result objects.  This
prevented the screen from refreshing when a custom module was
uninstalled.

I left in place the sql_upgrade_fx script which calls into the
SQLUpgradeService just to mitigate any potential code problems.
Eventually that file can be removed.

* Fix style problems