OpenEMR Portal and Main tabs render events
Added events to OpenEMR main tabs screen for outputting html at the end of the body tag and at the start of the body tag.
Added event to filter appointments in the portal home page.
Added event to render output at the end of the portal sections and before the scripts of the portal.
Refactored the appointments into a separate twig file that module writers can overwrite.