Patient portal appt fix for preferred categories.
If you had a category other than office_visit (catid=5) then a patient
couldn't find any provider available appointments to schedule with.  So
if additional categories were added etc, a patient could never make the
appointment.

Fixed this issue so now the finder will use the category for the search
just like it does with the provider.