Portal backport patch4 telehealthdeps (#17)
* Cherry Pick Calendar changes.

This is the result of cherry-picking this commit for the Calendar
extension points I did. https://github.com/openemr/openemr/commit/3e490e3d26ef9e9e9367ba3335517778a8604624

* Add in the global settings that was missing

* Twig module author extensions (#4715)

* Twig module author extensions

Added a twig extension to fire a generic event.  For now that is helpful
in being able to migrate over some of our php/html files that dispatch
render events.

I also added a twig environment event that module authors can hook into
in order to extend the file loader and load their own template files.  I
put an example in the twig environment event header of how to go about
doing this.

Also created some generic error handler pages with an example 400 page
so we can start having nicer looking error handling pages for when
things go wrong in OpenEMR.

* Fix spacing and imports

* Move patient date methods into service (#4716)

This came from a cherry-pick of https://github.com/openemr/openemr/commit/3d31e5512b5e56a8c225e1851188804d974a6567

Needed to leverage a number of patient date formatting methods which
were hard to get to w/ them being in the library.  Moved them into the
patient service method so we could access them from the service.

* Refactored session methods into classes (#4717)

Cherry picked from this commit f645f54368578f4a6369c3887fa354098eed242e

Moved the encounter and patient session setting functions from encounter
and pid.inc into classes so we can start using them anywhere we need w/
autoloading.

* clean up the uuid upgrade mechanism (#4546)

Came from cherry-pick of 7820a213e33d15db109cfb71ab6ef4e5171734bc

* fix upgrade version selected in dropdown and misc php8 fixes (#4629)

Note we only took the changes in sql_upgrade.php and did not bring in
changes to interface/patient_file/summary/dashboard_header.php as we
aren't handling that in this patch.

* fix upgrade version selected in dropdown and misc php8 fixes

* restore comment in docblock

* format advice

* move advice to paragraph and restore single xlt statement

* typo

* Openemr fix 4723 module install script uses OpenEMR sql upgrade service (#4724)

* Fix module semicolon install bug for custom module

Closes #4723. I migrate the logic in the sql_upgrade_fx file into a
SQLUpgradeService object that can be reused in other places.  I change
up the InstallerController for both the upgrade sql and the install sql
process to use the SQLUpgradeServiceObject.

I also fix a bug in the uninstall custom module that was throwing an
error due to a recent laminas pdo change with the result objects.  This
prevented the screen from refreshing when a custom module was
uninstalled.

I left in place the sql_upgrade_fx script which calls into the
SQLUpgradeService just to mitigate any potential code problems.
Eventually that file can be removed.

* Fix style problems

* Backport QueryUtils and sql.inc changes

Bringing in the query utils and the sql.inc changes.

* EventDispatcher hooks for patient create, and patient update (#4176)

cherry-picked from 38e524415d068b30d25d921c99fa403c14d01d9d

* Fixed issue in PatientService when passing array to addData(); Added escape_table to BaseService when getting uuid; Consolodated patient update and create to the PatientService;  Added events so modules can subscribe to patient update and create events; Added example module for usage of patient update/create hooks

* Added database insert functions to PatientService for use in core code, which trigger patient hooks

* revert ProcessingResult change that broke test

* make sure data parameters are in result

* fix type in function name mispelled databse

* Moved ref to PatientService to use statement

* Fix #4725 user events, uuid generation (#4726)

* Fix #4725 user events, uuid generation

Fixes #4725
Made it so the create user event only fires if the user successfully
saves.  Also made it so creating a user generates a uuid for it so we
have it at the time of the save user event.  Added some helper methods
to the user created and user updated event to make them more useful.

* Fix styles

* Add the inserted id to the patient event data (#4729)

* Backport PatientTrackerService and save event

* Backport system logger functions.

Need errorLogCaller functionality so we are backporting the system
logger.

* Fixing AppointmentService and backporting methods.

Backporting the methods we need to make the appointment service work for
module writers (such as telehealth).

We had to bring in the UuidRegistry changes to make both encounters,
users, and appts work properly.

Several methods were backported in the appointment service that relied
on newer FHIR functionality. Refactored the methods to use older
functionality so this will still work in 6_0_0.

Brought back the hasData() method in ProcessingResult (and other new
functions) to make this work.

* Fix space error in EncounterService

getEncountersBySearch breaks when you pass in search criteria due to an
extra space added in the method.

* Fix logger error.

Backported appointment service was relying on BaseService::getLogger()
which is not in version 6.0.0

Co-authored-by: Stephen Nielson <snielson@discoverandchange.com>