Fixes #4941 direct send with ccda
The direct send was not working with the CareCoordination module as it
had a duplicate code of the ccr/transmitCCD.  Working with EMRDirect
found that the Direct community expects one message per patient instead
of bundling multiple patients into a single message per recipient.  I
changed up the code flow to make this happen as well as removed the
duplicate code.  I was able to extend the ccr/transmitCCD.php file to
handle pdf and html files.

Also pulled out some of the constants in the transmitCCD function into
a class with better explanations.  This made the code more legible.

Also documented some of the AMC stuff with the care coordination.