Fixes #4888 Direct Message Whitelist
I discovered that the document white listing fails if the local
operating system's mime type detection on the document is
different than the mime type received via direct. This was
discovered with a CCDA xml file. EMRDirect was sending a mime type
of application/xml mime type and the local OS detected it as
text/xml. This breaks our current solution as we temporarily allow
the EMRDirect mime type if it matches the document mime type and
then we remove the temporary bypass.

Since we don't want to build a file quarantine process I went
with our original design approach and bypass the mime type check
altogether during the document storage process.
The files are protected on the drive as is so this should be fine.