CQRM Initial Cat1 with helper support.
Implemented the Date, View, and Cat1View helpers.  There are a couple of
other helpers they call out to which I've left unimplemented at this
point (search for TODO:)

Changed the mustache library to be the openemr mustache library that
injects the Context into the mustache template helper functions (on both
lambda's and object methods).