Add 3rd party api access url for portal.
Refactored the create credentials page to include the link to the FHIR
api and the fhir requirements doc.  I point the requirements doc
currently to our FHIR readme page but we may need this to be a dedicated
pdf or text document.  As part of this refactor I changed the templates
to be twig files.  Better separation of concerns and module writers can
now override the template.

Changed the size of the credentials popup to be a bit bigger to show the
larger amount of information we are displaying.