Fixes 4959, emails, amc patient access
So there is some additional work we'll need to do with the AMC rule
calculations for group methods but this commit adds in the FHIR 3rd
party api access instructions everywhere in the portal where accounts
are created / credentials generated to give patients access and
instructions on how to access their 3rd party apps.

I've also gone ahead and created twig email templates so that module
writers and others can overwrite the message emails if they so desire.

I added the Clinical Rule text descriptions to the AMC rules for the
patient access.  I also centralized the AMC titles/code data arrays into
the CertificationReportTypes class.