Openemr work on 4958 amc patient electronic access (#4997)
* WIP on #4959 - Add 2015 AMC support.

Added the 2015 AMC database fields to support 2015 rulesets.  Instead of
adding a bunch more branch conditions which complicates the code I
abstracted the certification report types into its own class that
contains the amc rule set type constants.

* Add default amc

On the patient dashboard it was breaking due to missing the default amc
piece.

* Add classes for AMC 315(g)(Required Test 2c)

Added in the classes and database properties for Required Test 2c.

* Refactored fhir url into its own class method.

Saw we were copying the fhir address url all over the place so I put it
into its own constant.

* Add 3rd party api access url for portal.

Refactored the create credentials page to include the link to the FHIR
api and the fhir requirements doc.  I point the requirements doc
currently to our FHIR readme page but we may need this to be a dedicated
pdf or text document.  As part of this refactor I changed the templates
to be twig files.  Better separation of concerns and module writers can
now override the template.

Changed the size of the credentials popup to be a bit bigger to show the
larger amount of information we are displaying.

* Fix undefined var error

* Style fixes

* Check in missing twig templates.

Missed the template twigs for the portal credentials refactor.

* Fixes 4959, emails, amc patient access

So there is some additional work we'll need to do with the AMC rule
calculations for group methods but this commit adds in the FHIR 3rd
party api access instructions everywhere in the portal where accounts
are created / credentials generated to give patients access and
instructions on how to access their 3rd party apps.

I've also gone ahead and created twig email templates so that module
writers and others can overwrite the message emails if they so desire.

I added the Clinical Rule text descriptions to the AMC rules for the
patient access.  I also centralized the AMC titles/code data arrays into
the CertificationReportTypes class.

* Fix logic error, database support, translations

* Fix escaping issue.

Co-authored-by: Stephen Nielson <snielson@discoverandchange.com>