Updated mustache templates to 3.1.8 qrda-reports
Found we were using a newer version of the mustache templates then what
we will test against for ONC.  I've gone ahead and changed those to the
old version for the Cat 3 reports.  Also found that the templates using
the ruby-ism of ? to denote a boolean question can't be processed in PHP.
So I've switched the templates that use the functionName? to drop the
question mark which allows those functions to actually execute
correctly.