QRDA Cat 3 Document Generation
Got the Cat 3 document to generate for a Cypress measure.

Had to heavily rework the report logic to more accurately match
what is  going on in the cypress cat 3 generation.  It a bit
convulated due to how the cqm-models, cypress, and cqm-reports
projects all interplay and convert data between the three but the
logic is now nearly identical as far as I can tell.

We still need to verify that stratified populations work correctly
as the measures I tested didn't appear to use that mechanism.  Also
apparently some measures can embed observations and additional
supplemental data into them so we need to have a way of testing
that also (unless our 10 measures chosen doesn't need any of that).

I need to download some of the population data and then compare it
against cypress and make sure that all is working correctly.