Qdm 2022 qrda export (#5046)
* Added 5.6 QDM models, and a readme how to generate new models

* merge date helper

* Intermediate

* intermediate

* ported a bunch of Ruby code to support Cat III to PHP

* intermediate

* finished porting coding only, haven't run/tested

* Hook up basic rendering of Cat3 XML

* intermediate, walking through template population

* intermediate

* intermedite: added in result calculator

* fix input to export service, now just requires array of measure strings

* added patient addresses and telcoms for qrda cat 1 export

* Updated mustache templates to 3.1.8 qrda-reports

Found we were using a newer version of the mustache templates then what
we will test against for ONC.  I've gone ahead and changed those to the
old version for the Cat 3 reports.  Also found that the templates using
the ruby-ism of ? to denote a boolean question can't be processed in PHP.
So I've switched the templates that use the functionName? to drop the
question mark which allows those functions to actually execute
correctly.

* Updated to latest oe-cqm-module version.

* resolve patient extension

* fix date format for JSON cqm calculation, add result to procedure QDM

* fix typos in query

* fix styling

* trying to fix styles again

* try fixing style again

Co-authored-by: Stephen Nielson <snielson@discoverandchange.com>