WIP #4959 Rule def,help files, db changes
Created and implemented the AMC 315g Required Rule 7 which is the Support
Electronic Referral Loops by Sending Health Information rule for MIPS.

Updated the transactions help file to include the encounter id
information and MIPS reporting pieces.  Implemented a help file for the
AMC report that also explains how to run the report and specific
information for MIPS calculations.

Added a new checkbox flag on the transactions for providers to confirm
that the recipient they sent a summary of care document to actually
received the document.

Made it so the confirmation receipt is set to true when EMRDirect is
sent since EMRDirect automatically confirms that the recipient received
the message.

Added the new rule definitions to the database.