Fixes #4959 Referral Loops
Changed the AMC reporting for sending the referral loop with care
coordination.  Reworked the tracking of problems, medications, and
allergies to be an AND clause instead of an OR condition for reporting.
Made it so the ccda is directly connected via the database to a
transaction referral.  Changed the 315(g)(2) Test 7 implementation to
use the amc's ccda valid flag.

This fully implements #4959 referral loops for the test.