AMC Fix denominator calculation.
When we added the inclusion of referrals into the buildPatientData array
it makes the 2c tests fail since patients were included that had a
referral to a provider (needed for AMC test 7) but had no encounter.
Since the denominator rule is that the patient has to have been seen in
an encounter check that in the denominator.

Currently we don't specify explicit action labels for the denominator
but I built some initial support if ONC ends up mandating we have to
have that.