Changed the patient details report for printing.
The printing on individual reports was overlaying table headers since we
had a table nested inside a table and the browser printer was getting
confused.

I rewrote the report to drop the outer table and make it so we had a
single tabulated table that made the print work out properly.