Cleaned up cqm report page
Removed a lot of the old unused amc options but made it so old reports
still display properly.  Migrated the entire cqm.php file over to php
and added twig extensions for display help icons and the oemrUI below
container extension.  Also added extensions for dealing with the wierd
date settings that are used in some of the twig functions.