CQM Matched Patient Ids with calculator
Tacoma project CQM Calculator uses Mongoose ObjectId for its id
properties. We can't just pass straight through the patient id
in order to get the matched patient ids back.  Instead we have to
pass in a stringified format for the patient id that matches
the serialized BSON format for the ObjectId.  In this case its a
24 character string in hexidecimal format that represents the pid.

Fortunately the bigint data type is only 8 bytes and ObjectId is a 12
byte format so it fits within the serialization.