Fixed bug in supplemental data
Identified and fixed a bug in the supplemental data for the cat 3 report
calculator. Made the supplemental patient data use code values of 1
until @kchapple's work gets in to fix the QDM queries on the patient.

Put in a comment about the provider definition for the CQM commenting
with a link to the 2022 implementation guide for CQM.