Openemr cqm supplemental data (#5113)
* CQM Matched Patient Ids with calculator

Tacoma project CQM Calculator uses Mongoose ObjectId for its id
properties. We can't just pass straight through the patient id
in order to get the matched patient ids back.  Instead we have to
pass in a stringified format for the patient id that matches
the serialized BSON format for the ObjectId.  In this case its a
24 character string in hexidecimal format that represents the pid.

Fortunately the bigint data type is only 8 bytes and ObjectId is a 12
byte format so it fits within the serialization.

* Fixed bug in supplemental data

Identified and fixed a bug in the supplemental data for the cat 3 report
calculator. Made the supplemental patient data use code values of 1
until @kchapple's work gets in to fix the QDM queries on the patient.

Put in a comment about the provider definition for the CQM commenting
with a link to the 2022 implementation guide for CQM.

* Map results into supplememntal counts for rendering for race, ethnicity, sex, etc

* Fix styles

* fix reporting year and submission program, and attempt to not render payer (cypress still doesn't like)

* remove unused fn

* add missing care plan service to builder

* fix for determinig existing populations

* fix styles

Co-authored-by: Stephen Nielson <snielson@discoverandchange.com>