WIP fix patient scopes when used by users
Need to handle in the standalone application context the ability to use
patient scopes while logged in as a practitioner.  We originally
restricted patient scopes to JUST patients and user scopes to JUST
users.  However, latest inferno and v2 of smart on FHIR is expecting a
practitioner to be able to select a user and have their scopes
restricted to JUST that user.

Currently I have this working when used for a single patient.  However,
because the trusted user session is re-used across user logins for a
single client, the bound patient for the access token is updated across
all access tokens regardless of the user that was chosen at the time of
the access token creation.  I haven't decided yet on whether this is a
problem or not.  It very much violates the element of least surprise
principle and could have unintended side effects where the provider
authorizes an app in one location with one patient and then authorizes a
different bound patient for the same app client in a different window
, only to have the information in the new client stomp on the old.

I think the only way to get away from this is either bind up the patient
that is in the selected context into the access token or perhaps have a
trusted user session per access token. I'll have to think on this more
tomorrow.