Fix shared access token session context
This fix makes it so that each access token has a uniquely bound session
context rather than using a shared session context that was populated from
the trusted user session cache.  Now SMART applications can have different
patients open using different access tokens for the same application without
stomping on each other's data.

Modified refresh token grant so we can pull our access token context and
repopulate it when a new access token is issued from the refresh token.

Refactored IdTokenSmartResponse to pull the Access Token response context
variables into its own class called SMARTSessionTokenContextBuilder that is used
to build the context information.  This is needed to deal with patient bound
contexts and contexts that are re-issued via refresh tokens.

Changed up AccessTokenRepository to save off the access token context
information (such as a selected patient, encounter, etc) that needs to be
retrieved / filtered for each access token (or reissued when a refresh token
generates a new access token).

Changed up AuthorizationController,SMARTAuthorizationController to set the
session pid to be the actual patient pid and not the uuid of the patient.
Added the puuid to the session.

Modified dispatch.php to call RestConfig to populate the access token context
instead of grabbing it from the trusted_user session_cache.