Fix patient scopes when used by users (#5151)
Fix patient scopes when used by users

Need to handle in the standalone application context the ability to use
patient scopes while logged in as a practitioner.  We originally
restricted patient scopes to JUST patients and user scopes to JUST
users.  However, latest inferno and v2 of smart on FHIR is expecting a
practitioner to be able to select a user and have their scopes
restricted to JUST that user.

Currently I have this working when used for a single patient.  However,
because the trusted user session is re-used across user logins for a
single client, the bound patient for the access token is updated across
all access tokens regardless of the user that was chosen at the time of
the access token creation.  I haven't decided yet on whether this is a
problem or not.  It very much violates the element of least surprise
principle and could have unintended side effects where the provider
authorizes an app in one location with one patient and then authorizes a
different bound patient for the same app client in a different window
, only to have the information in the new client stomp on the old.

I think the only way to get away from this is either bind up the patient
that is in the selected context into the access token or perhaps have a
trusted user session per access token. I'll have to think on this more
tomorrow.

* Fix shared access token session context

This fix makes it so that each access token has a uniquely bound session
context rather than using a shared session context that was populated from
the trusted user session cache.  Now SMART applications can have different
patients open using different access tokens for the same application without
stomping on each other's data.

Modified refresh token grant so we can pull our access token context and
repopulate it when a new access token is issued from the refresh token.

Refactored IdTokenSmartResponse to pull the Access Token response context
variables into its own class called SMARTSessionTokenContextBuilder that is used
to build the context information.  This is needed to deal with patient bound
contexts and contexts that are re-issued via refresh tokens.

Changed up AccessTokenRepository to save off the access token context
information (such as a selected patient, encounter, etc) that needs to be
retrieved / filtered for each access token (or reissued when a refresh token
generates a new access token).

Changed up AuthorizationController,SMARTAuthorizationController to set the
session pid to be the actual patient pid and not the uuid of the patient.
Added the puuid to the session.

Modified dispatch.php to call RestConfig to populate the access token context
instead of grabbing it from the trusted_user session_cache.

* Fix styles