Merge FHIR Bug Fixes To Release 6.1.0 (#5160)
* Fixes #5095 Patient,Goals,Immunization (#5097)

Not sure what happened but it looks like provider_uuid got stomped on in
CarePlan and Immunization.  Adding that in to our uuid mappings fixes
the problem with the resources returning 500.

For Patient it looks like the patient/* context was broken in the
Patient resource endpoint when we refactored the parameter name for
documentation purposes.  I've updated the $id parameter in the patient
route to be $uuid.

* Openemr fhir single patient api #5122 (#5129)

* Fix SurgeryService query.

FHIR searchAll was relying on a username column that doesn't exist.  Not
sure if some code got stomped on or how this ever passed our inferno
test suite.

* Fix Provenance queries.

Again not sure if code got stomped on but DiagnosticReports and
DocumentReference and Immunization Provenance were not working with
single patient api.

* FHIR document how medication works inside request

Added comments that explain how medication is embedded via the code
property inside of MedicationRequest.

* Fixes #5125,#5123 FHIR ids,Provenance

Reworked the FHIR ids for Provenance,CarePlan, and Goals.  FHIR R4.0.1
changed the logical id allowed fields (or at least inferno started
validating them) so we had to adjust how we handle these resources.  I
use a different surrogate key string which is valid in 4.0.1.  I also
handle the prior key data so we can still read and return the old format
so people's data doesn't break.

I also fix the issue where clinical notes,procedures and immunization weren't
returning provenance resources

* Fix style issues.

* Fix Care Plan unit tests.

We had too many fixture interdependencies now that the careplan also
joins on the forms table. Since I was testing the
surrogate key logic I stopped the getOne test and just focused on the
surrogateKey.

* Fixes #5137 Vitals save data absent value (#5138)

Make it so we can save a 0 value for columns in vitals to signify the
data is missing.

* Fixes #5127 mapped uuids not saving in registry (#5139)

Made it so the uuids for mapped resources such as vitals are being
populated in the uuid registry.  Users will need to run the
sql_upgrade.php for any existing installation in order for the vitals to
get their registry populated.