Fixes #5181 CareTeams,Minor FHIR bug fixes
Changed up the care team value sets since we can't use NUCC anymore and
instead need to use a SNOMED-CT valuset for the care team provider.
Oddly there is no 'valid' value for facilities as the role is all
provider centric.  However, SNOMED-CT still has values for health care
organization so we use that for our code.

Added some fixes to get rid of warnings with the smart session context
builder.

Also Changed up the group export to use getOne which is what we really
want instead of hitting getAll.