Add 2 digit language codes for FHIR (#5182)
Fixes #5183 
* Add 2 digit language codes for FHIR

Add in the 2 digit language codes to the codes column so we can pass
ccda and FHIR validation. 2 digit is required to be used if one
available otherwise we use 3 digit language codes.

* Language fixes - notes column

Don't use codes column for language, keep with  notes column and drop 3
digit language where we don't need them.