Fixes #5190 DocumentReference categories
I have added a codes column to the category table for documents so you
can set a specific LOINC document category code to a documents category.

The LOINC codes come from https://loinc.org/document-ontology/current-version/

I made it so the codes column is optional.  I also fixed up a bit of the
CategoryTree code to decouple itself from the CategoryTree class.

I set a couple of different LOINC codes for categories.  There are more
that fit but this at least gets us through MU3 validation.

I had to completely drop the openemr type and instead report a Data
Absent Reason if a document category doesn't have a LOINC code.

In order to get the codes selector to work I switched the document
categories over to a twig file.  Couldn't get the javascript to execute
properly in smarty and it was a simple enough file to port.  So one less
smarty file to deal with.