Fixes #5201 Report US Core Profiles for capability
Inferno validator changed up some of the us core profiles we had to
report.  We also were missing a bunch of them in other resources.  This
gets rid of the inferno warning on our capability statement.  It also
advertises to FHIR clients that we support the full US Core profile
spectrum.