Fixes #5215 provenance algorithm corrections
Made the provenance use its own logic instead of relying on the bulk
export trait.  We first only grab resources that have been included in
the JOB (which are filtered by scopes and patient queries).  Then we
correctly set the patient filters on the individual sub resources.
Lastly instead of wrapping everything up in a processing result which is
memory and cpu intensive, we write things out to the stream as we go
through the system.