Fixes #5220 Send Trusted Direct Message
Implemented the ability to send messages w/o attachments via the EMR
Direct service over the Direct protocol.

Built the ability to select a document that has been uploaded to the
system and send it via EMRDirect.

Built a document finder dialog window so that you can select a patient
or non-patient document using the finder.  It checks against the current
user's ACLs and the document category's acos to determine if the user
can see the document.  It only shows 25 results at a time and requires
the user to be more specific in their search if they want to see more.
By default it displays the most recent documents uploaded.