Implemented Document File Preview for messages.
Messages was throwing a javascript error if you didn't assign a patient
for a document.  However, often the document is what contains the
information about what patient the file should belong to.  I made it so
that if there is no assigned patient you can at least view the document
in the Miscellaneous -> New Documents screen.

This also facilitates testing with ONC as we have to demonstrate that
the Direct file was successfully decrypted and viewable.