Fixes #5255 observation forms reason codes
Added support for reason code, reason_status, and reason_description for
the observation form as well as for the vitals_form.

Fixed the observation form to use a datetime selector instead of just a
regular date field.

Implemented a reason code widget script that allows us to implement the
code selector and hide/display of the reason container based upon the
launch button.

Refactored a lot of the javascript i the observation and vitals form
into separate javascript files to simplify debugging and to make it
easier to work with.

Added reason_code, reason_status, and reason_description to the
vital_form_details table.

Added a callback function parameter to the dialog so that you can pass a
callback that doesn't have to exist in the global document window.