Add Reason support to careplan Fixes #5267
Added the reason code and start and end date selectors for the care plan
entries.  Also added the ability to have more code type selectors in the
reason code popup.  Added comments to explain the low and high on the
reason code in care plan.

Also added a comment in FHIR to explain how we'd leverage reason code if
we start reporting on the overall activity for CarePlan.