WIP encounter type field in encounter form.
Since we can't bundle certain codes like CPT4 or even SNOMED in
non-member countries we want to only install the codes property for
the Encounter Type list options if we have the right code systems
installed.  So we have added a CodeTypes event subscriber that listens
to newly installed coded types as well as the SQL Upgrade script process
to update the Encounter Type codes property.

This is WIP and is not finalized.