WIP #5277 Immunization reason codes
This depends on the encounter_type work so we are making this a WIP for
now.  We need to handle both the reason and negationReason which for
immunization is the refusal_reason code.  I couldn't find any examples
where both a reason and negationRationale QDM exist simultaneously but
the spec supports the use case.  Since we already had reason codes
populated I map the valuesets for QDM to the NIP002 CDC code mappings in
the SQL so we can still support reporting to immunization registries.

I also updated the FHIR to support the religious objection since that's
the only additional value code that makes sense in addition to patient
objection for what FHIR allows.  I imagine future versions of FHIR will
be more in line with CCDA/QRDA for the immunization reason codes.